+++ aiter/ops/triton/_triton_kernels/attention/deepseek_sparse_attention.py
@@ -22,6 +22,28 @@ import torch
 import triton
 import triton.language as tl
 
+from ._dsa_bwd_preprocess import _sparse_mla_bwd_preprocess
+from ._dsa_bwd_fused import _sparse_mla_bwd_kernel
+from ._dsa_bwd_recompute import _bwd_dq_only, _bwd_dkv_hg_fused_recompute
+from ._dsa_bwd_split_intermediate import _bwd_dq_store_intermediates, _bwd_dkv_hg_fused
+from ._dsa_bwd_privatized import (
+    _bwd_dkv_privatized,
+    _bwd_dkv_xcd_local,
+    _bwd_dkv_nonatomic_scatter,
+    _bwd_dkv_reduce_copies,
+)
+from ._dsa_bwd_gather import (
+    _bwd_compute_dkv_intermediate,
+    _bwd_dkv_gather,
+    _build_inverted_topk,
+    _build_inverted_topk_slice,
+    _bwd_chunk_dq_store_ds,
+    _bwd_chunk_dq,
+    _bwd_chunk_dkv_interm,
+    _bwd_dkv_gather_acc,
+)
+from ._dsa_bwd_persistent import _bwd_persistent_chunk, _bwd_chunk_reduce
+
 
 # =====================================================================
 # Utility
@@ -203,1665 +225,6 @@ def _sparse_mla_fwd_train_kernel(
         LSE_ptr + token_idx * num_heads + offs_h,
         lse, mask=mask_h,
     )
-
-
-# =====================================================================
-# Backward — preprocess kernel (Delta computation)
-# =====================================================================
-@triton.jit
-def _sparse_mla_bwd_preprocess(
-    O_ptr,          # [total_tokens, num_heads, D_V]
-    dO_ptr,         # [total_tokens, num_heads, D_V]
-    Delta_ptr,      # [total_tokens, num_heads]
-    stride_o_t: tl.int64,
-    stride_o_h: tl.int64,
-    num_heads: tl.int32,
-    D_V: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-):
-    """
-    Delta[t, h] = sum_d(O[t, h, d] * dO[t, h, d])
-
-    Grid: (total_tokens, cdiv(num_heads, BLOCK_H))
-    """
-    token_idx = tl.program_id(0)
-    hg_idx = tl.program_id(1)
-
-    offs_h = hg_idx * BLOCK_H + tl.arange(0, BLOCK_H)
-    mask_h = offs_h < num_heads
-    offs_d = tl.arange(0, D_V)
-
-    base = token_idx * stride_o_t
-
-    O = tl.load(
-        O_ptr + base + offs_h[:, None] * stride_o_h + offs_d[None, :],
-        mask=mask_h[:, None], other=0.0,
-    )
-    dO = tl.load(
-        dO_ptr + base + offs_h[:, None] * stride_o_h + offs_d[None, :],
-        mask=mask_h[:, None], other=0.0,
-    )
-
-    delta = tl.sum(O.to(tl.float32) * dO.to(tl.float32), axis=1)
-
-    tl.store(
-        Delta_ptr + token_idx * num_heads + offs_h,
-        delta, mask=mask_h,
-    )
-
-
-# =====================================================================
-# Backward — autotune configs (for fused baseline)
-# =====================================================================
-def _bwd_prune_configs(configs, named_args, **kwargs):
-    """Prune autotune configs that would exceed per-CU LDS or hit known bugs."""
-    D_V = kwargs.get("D_V", named_args.get("D_V"))
-    D_ROPE = kwargs.get("D_ROPE", named_args.get("D_ROPE"))
-    pruned = []
-    for config in configs:
-        bh = config.kwargs["BLOCK_H"]
-        tk = config.kwargs["TILE_K"]
-        ns = config.num_stages
-        kv_lds = (D_V + D_ROPE) * tk * 2 * ns
-        if kv_lds > _LDS_LIMIT:
-            continue
-        # Skip BLOCK_H=64 / TILE_K=16 / num_warps=4 / num_stages=1:
-        # produces NaN on AMD CDNA due to compiler bug.
-        if bh == 64 and tk == 16 and config.num_warps == 4 and ns == 1:
-            continue
-        pruned.append(config)
-    if not pruned:
-        pruned.append(configs[0])
-    return pruned
-
-
-def _get_bwd_autotune_configs():
-    configs = []
-    for BLOCK_H in [16, 32, 64]:
-        for TILE_K in [16, 32, 64, 128]:
-            for num_warps in [2, 4, 8, 16]:
-                for num_stages in [1, 2, 3, 4]:
-                    configs.append(
-                        triton.Config(
-                            {"BLOCK_H": BLOCK_H, "TILE_K": TILE_K},
-                            num_warps=num_warps,
-                            num_stages=num_stages,
-                        )
-                    )
-    return configs
-
-
-# =====================================================================
-# Backward method="fused" — single fused kernel (baseline, 58ms)
-# =====================================================================
-@triton.autotune(
-    configs=_get_bwd_autotune_configs(),
-    key=["num_heads", "TOPK", "D_V", "D_ROPE"],
-    prune_configs_by={"early_config_prune": _bwd_prune_configs},
-    reset_to_zero=["dKV_ptr"],
-)
-@triton.jit
-def _sparse_mla_bwd_kernel(
-    Q_ptr,          # [total_tokens, num_heads, D_QK]
-    KV_ptr,         # [total_tokens, 1, D_QK]
-    dO_ptr,         # [total_tokens, num_heads, D_V]
-    TopK_ptr,       # [total_tokens, topk]
-    LSE_ptr,        # [total_tokens, num_heads]  float32
-    Delta_ptr,      # [total_tokens, num_heads]  float32
-    dQ_ptr,         # [total_tokens, num_heads, D_QK]
-    dKV_ptr,        # [total_tokens, D_QK]  float32 (atomic target, squeezed)
-    Q_T_ptr,        # [total_tokens, D_QK, num_heads]
-    dO_T_ptr,       # [total_tokens, D_V, num_heads]
-    stride_q_t: tl.int64,
-    stride_q_h: tl.int64,
-    stride_kv_t: tl.int64,
-    stride_do_t: tl.int64,
-    stride_do_h: tl.int64,
-    stride_dq_t: tl.int64,
-    stride_dq_h: tl.int64,
-    stride_dkv_t: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_qt_t: tl.int64,
-    stride_dot_t: tl.int64,
-    scale: tl.float32,
-    num_heads: tl.int32,
-    TOPK: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-    TILE_K: tl.constexpr,
-    D_V: tl.constexpr,
-    D_ROPE: tl.constexpr,
-):
-    """
-    Fused sparse MLA backward: dQ + dKV.
-
-    Grid: (total_tokens, cdiv(num_heads, BLOCK_H))
-    Each program: 1 query token x BLOCK_H heads.
-    """
-    token_idx = tl.program_id(0)
-    hg_idx = tl.program_id(1)
-
-    offs_h = hg_idx * BLOCK_H + tl.arange(0, BLOCK_H)
-    mask_h = offs_h < num_heads
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    q_base = token_idx * stride_q_t
-    Q_lora = tl.load(
-        Q_ptr + q_base + offs_h[:, None] * stride_q_h + offs_v[None, :],
-        mask=mask_h[:, None], other=0.0,
-    )
-    Q_rope = tl.load(
-        Q_ptr + q_base + offs_h[:, None] * stride_q_h + (D_V + offs_r[None, :]),
-        mask=mask_h[:, None], other=0.0,
-    )
-
-    do_base = token_idx * stride_do_t
-    dO_val = tl.load(
-        dO_ptr + do_base + offs_h[:, None] * stride_do_h + offs_v[None, :],
-        mask=mask_h[:, None], other=0.0,
-    )
-
-    qt_base = token_idx * stride_qt_t
-    Q_lora_T = tl.load(
-        Q_T_ptr + qt_base + offs_v[:, None] * num_heads + offs_h[None, :],
-        mask=mask_h[None, :], other=0.0,
-    )
-    Q_rope_T = tl.load(
-        Q_T_ptr + qt_base + (D_V + offs_r[:, None]) * num_heads + offs_h[None, :],
-        mask=mask_h[None, :], other=0.0,
-    )
-
-    dot_base = token_idx * stride_dot_t
-    dO_T = tl.load(
-        dO_T_ptr + dot_base + offs_v[:, None] * num_heads + offs_h[None, :],
-        mask=mask_h[None, :], other=0.0,
-    )
-
-    lse = tl.load(
-        LSE_ptr + token_idx * num_heads + offs_h,
-        mask=mask_h, other=0.0,
-    )
-    delta = tl.load(
-        Delta_ptr + token_idx * num_heads + offs_h,
-        mask=mask_h, other=0.0,
-    )
-
-    dQ_lora = tl.zeros([BLOCK_H, D_V], dtype=tl.float32)
-    dQ_rope = tl.zeros([BLOCK_H, D_ROPE], dtype=tl.float32)
-
-    NUM_TILES: tl.constexpr = (TOPK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t
-    offs_tile = tl.arange(0, TILE_K)
-
-    topk_pos = tl.load(
-        TopK_ptr + topk_base + offs_tile,
-        mask=offs_tile < TOPK, other=-1,
-    )
-    topk_pos_next = topk_pos
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        valid = (tile_start + offs_tile) < TOPK
-        valid = valid & (topk_pos != -1)
-
-        if t + 1 < NUM_TILES:
-            next_offs = (t + 1) * TILE_K + offs_tile
-            topk_pos_next = tl.load(
-                TopK_ptr + topk_base + next_offs,
-                mask=next_offs < TOPK, other=-1,
-            )
-
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        K_lora_T = tl.load(
-            KV_ptr + safe_pos[None, :] * stride_kv_t + offs_v[:, None],
-            mask=valid[None, :], other=0.0,
-        )
-        K_rope_T = tl.load(
-            KV_ptr + safe_pos[None, :] * stride_kv_t + (D_V + offs_r[:, None]),
-            mask=valid[None, :], other=0.0,
-        )
-
-        S = tl.dot(Q_lora, K_lora_T)
-        S += tl.dot(Q_rope, K_rope_T)
-        S *= scale
-        S = tl.where(valid[None, :] & mask_h[:, None], S, float("-inf"))
-
-        P = tl.exp(S - lse[:, None])
-        P = tl.where(valid[None, :] & mask_h[:, None], P, 0.0)
-
-        dP = tl.dot(dO_val, K_lora_T)
-        dS = P * (dP - delta[:, None]) * scale
-        dS = tl.where(valid[None, :] & mask_h[:, None], dS, 0.0)
-
-        V_lora = tl.trans(K_lora_T)
-        dQ_lora += tl.dot(dS.to(V_lora.dtype), V_lora).to(tl.float32)
-
-        K_rope = tl.trans(K_rope_T)
-        dQ_rope += tl.dot(dS.to(K_rope.dtype), K_rope).to(tl.float32)
-
-        dKV_lora_T = tl.dot(Q_lora_T, dS.to(Q_lora_T.dtype))
-        dKV_lora_T += tl.dot(dO_T, P.to(dO_T.dtype))
-        dKV_lora_T = dKV_lora_T.to(tl.float32)
-
-        dKV_rope_T = tl.dot(Q_rope_T, dS.to(Q_rope_T.dtype))
-        dKV_rope_T = dKV_rope_T.to(tl.float32)
-
-        dkv_ptrs_lora = (
-            dKV_ptr + safe_pos[None, :] * stride_dkv_t + offs_v[:, None]
-        )
-        tl.atomic_add(dkv_ptrs_lora, dKV_lora_T, mask=valid[None, :], sem="relaxed")
-
-        dkv_ptrs_rope = (
-            dKV_ptr + safe_pos[None, :] * stride_dkv_t + (D_V + offs_r[:, None])
-        )
-        tl.atomic_add(dkv_ptrs_rope, dKV_rope_T, mask=valid[None, :], sem="relaxed")
-
-        if t + 1 < NUM_TILES:
-            topk_pos = topk_pos_next
-
-    dq_base = token_idx * stride_dq_t
-    tl.store(
-        dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + offs_v[None, :],
-        dQ_lora.to(Q_lora.dtype), mask=mask_h[:, None],
-    )
-    tl.store(
-        dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + (D_V + offs_r[None, :]),
-        dQ_rope.to(Q_rope.dtype), mask=mask_h[:, None],
-    )
-
-
-# =====================================================================
-# Backward method="recompute" — dQ kernel (no intermediate stores)
-# =====================================================================
-@triton.jit
-def _bwd_dq_only(
-    Q_ptr, KV_ptr, dO_ptr, TopK_ptr, LSE_ptr, Delta_ptr,
-    dQ_ptr,
-    stride_q_t: tl.int64, stride_q_h: tl.int64,
-    stride_kv_t: tl.int64,
-    stride_do_t: tl.int64, stride_do_h: tl.int64,
-    stride_dq_t: tl.int64, stride_dq_h: tl.int64,
-    stride_topk_t: tl.int64,
-    scale: tl.float32, num_heads: tl.int32,
-    TOPK: tl.constexpr, BLOCK_H: tl.constexpr, TILE_K: tl.constexpr,
-    D_V: tl.constexpr, D_ROPE: tl.constexpr,
-):
-    """Pure dQ kernel -- computes dots 1-5, stores only dQ. No intermediates."""
-    token_idx = tl.program_id(0)
-    hg_idx = tl.program_id(1)
-    offs_h = hg_idx * BLOCK_H + tl.arange(0, BLOCK_H)
-    mask_h = offs_h < num_heads
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    q_base = token_idx * stride_q_t
-    Q_lora = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + offs_v[None, :],
-                     mask=mask_h[:, None], other=0.0)
-    Q_rope = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + (D_V + offs_r[None, :]),
-                     mask=mask_h[:, None], other=0.0)
-    do_base = token_idx * stride_do_t
-    dO_val = tl.load(dO_ptr + do_base + offs_h[:, None] * stride_do_h + offs_v[None, :],
-                     mask=mask_h[:, None], other=0.0)
-    lse = tl.load(LSE_ptr + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-    delta = tl.load(Delta_ptr + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-
-    dQ_lora = tl.zeros([BLOCK_H, D_V], dtype=tl.float32)
-    dQ_rope = tl.zeros([BLOCK_H, D_ROPE], dtype=tl.float32)
-    NUM_TILES: tl.constexpr = (TOPK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t
-    offs_tile = tl.arange(0, TILE_K)
-    topk_pos = tl.load(TopK_ptr + topk_base + offs_tile, mask=offs_tile < TOPK, other=-1)
-    topk_pos_next = topk_pos
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        valid = (tile_start + offs_tile) < TOPK
-        valid = valid & (topk_pos != -1)
-        if t + 1 < NUM_TILES:
-            next_offs = (t + 1) * TILE_K + offs_tile
-            topk_pos_next = tl.load(TopK_ptr + topk_base + next_offs,
-                                    mask=next_offs < TOPK, other=-1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        K_lora_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + offs_v[:, None],
-                          mask=valid[None, :], other=0.0)
-        K_rope_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + (D_V + offs_r[:, None]),
-                          mask=valid[None, :], other=0.0)
-
-        S = tl.dot(Q_lora, K_lora_T)
-        S += tl.dot(Q_rope, K_rope_T)
-        S *= scale
-        S = tl.where(valid[None, :] & mask_h[:, None], S, float("-inf"))
-        P = tl.exp(S - lse[:, None])
-        P = tl.where(valid[None, :] & mask_h[:, None], P, 0.0)
-
-        dP = tl.dot(dO_val, K_lora_T)
-        dS = P * (dP - delta[:, None]) * scale
-        dS = tl.where(valid[None, :] & mask_h[:, None], dS, 0.0)
-
-        V_lora = tl.trans(K_lora_T)
-        dQ_lora += tl.dot(dS.to(V_lora.dtype), V_lora).to(tl.float32)
-        K_rope = tl.trans(K_rope_T)
-        dQ_rope += tl.dot(dS.to(K_rope.dtype), K_rope).to(tl.float32)
-
-        if t + 1 < NUM_TILES:
-            topk_pos = topk_pos_next
-
-    dq_base = token_idx * stride_dq_t
-    tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + offs_v[None, :],
-             dQ_lora.to(Q_lora.dtype), mask=mask_h[:, None])
-    tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + (D_V + offs_r[None, :]),
-             dQ_rope.to(Q_rope.dtype), mask=mask_h[:, None])
-
-
-# =====================================================================
-# Backward method="recompute" — dKV kernel with full recomputation
-# =====================================================================
-@triton.jit
-def _bwd_dkv_hg_fused_recompute(
-    Q_ptr,          # [T, H, D_QK] bf16
-    KV_ptr,         # [T, 1, D_QK] bf16
-    dO_ptr,         # [T, H, D_V] bf16
-    Q_T_ptr,        # [T, D_QK, H] bf16
-    dO_T_ptr,       # [T, D_V, H] bf16
-    TopK_ptr,       # [T, TOPK] int32
-    LSE_ptr,        # [T, H] fp32
-    Delta_ptr,      # [T, H] fp32
-    dKV_ptr,        # [T, D_QK] fp32
-    stride_q_t: tl.int64, stride_q_h: tl.int64,
-    stride_kv_t: tl.int64,
-    stride_do_t: tl.int64, stride_do_h: tl.int64,
-    stride_qt_t: tl.int64,
-    stride_dot_t: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_dkv_t: tl.int64,
-    scale: tl.float32, num_heads: tl.int32,
-    TOPK: tl.constexpr,
-    TILE_K: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-    NUM_HG: tl.constexpr,
-    D_V: tl.constexpr,
-    D_ROPE: tl.constexpr,
-):
-    """
-    dKV kernel with full recomputation -- NO intermediate buffers needed.
-
-    Grid: (total_tokens,) -- ONE program per token.
-    Each program: loads Q, dO, lse, delta per HG, recomputes S, P, dS,
-    then computes dKV (dots 6-8). Scatters dKV once per tile.
-    """
-    token_idx = tl.program_id(0)
-
-    NUM_TILES: tl.constexpr = (TOPK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t
-    offs_tile = tl.arange(0, TILE_K)
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        tile_offs = tile_start + offs_tile
-        topk_pos = tl.load(TopK_ptr + topk_base + tile_offs,
-                           mask=tile_offs < TOPK, other=-1)
-        valid = (tile_offs < TOPK) & (topk_pos != -1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        K_lora_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + offs_v[:, None],
-                          mask=valid[None, :], other=0.0)
-        K_rope_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + (D_V + offs_r[:, None]),
-                          mask=valid[None, :], other=0.0)
-
-        dKV_lora = tl.zeros([D_V, TILE_K], dtype=tl.float32)
-        dKV_rope = tl.zeros([D_ROPE, TILE_K], dtype=tl.float32)
-
-        for hg in range(NUM_HG):
-            offs_h = hg * BLOCK_H + tl.arange(0, BLOCK_H)
-            mask_h = offs_h < num_heads
-
-            q_base = token_idx * stride_q_t
-            Q_lora = tl.load(
-                Q_ptr + q_base + offs_h[:, None] * stride_q_h + offs_v[None, :],
-                mask=mask_h[:, None], other=0.0,
-            )
-            Q_rope = tl.load(
-                Q_ptr + q_base + offs_h[:, None] * stride_q_h + (D_V + offs_r[None, :]),
-                mask=mask_h[:, None], other=0.0,
-            )
-            dO_val = tl.load(
-                dO_ptr + token_idx * stride_do_t + offs_h[:, None] * stride_do_h + offs_v[None, :],
-                mask=mask_h[:, None], other=0.0,
-            )
-
-            lse = tl.load(LSE_ptr + token_idx * num_heads + offs_h,
-                         mask=mask_h, other=0.0)
-            delta_val = tl.load(Delta_ptr + token_idx * num_heads + offs_h,
-                               mask=mask_h, other=0.0)
-
-            # Recompute S, P, dS (dots 1-3)
-            S = tl.dot(Q_lora, K_lora_T)
-            S += tl.dot(Q_rope, K_rope_T)
-            S *= scale
-            S = tl.where(valid[None, :] & mask_h[:, None], S, float("-inf"))
-            P = tl.exp(S - lse[:, None])
-            P = tl.where(valid[None, :] & mask_h[:, None], P, 0.0)
-
-            dP = tl.dot(dO_val, K_lora_T)
-            dS_val = P * (dP - delta_val[:, None]) * scale
-            dS_val = tl.where(valid[None, :] & mask_h[:, None], dS_val, 0.0)
-
-            # Load Q_T, dO_T for dKV (dots 6-8)
-            qt_base = token_idx * stride_qt_t
-            Q_lora_T = tl.load(
-                Q_T_ptr + qt_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            Q_rope_T = tl.load(
-                Q_T_ptr + qt_base + (D_V + offs_r[:, None]) * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            dot_base = token_idx * stride_dot_t
-            dO_T = tl.load(
-                dO_T_ptr + dot_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-
-            dKV_lora += tl.dot(Q_lora_T, dS_val.to(Q_lora_T.dtype)).to(tl.float32)
-            dKV_lora += tl.dot(dO_T, P.to(dO_T.dtype)).to(tl.float32)
-            dKV_rope += tl.dot(Q_rope_T, dS_val.to(Q_rope_T.dtype)).to(tl.float32)
-
-        dkv_ptrs_lora = dKV_ptr + safe_pos[None, :] * stride_dkv_t + offs_v[:, None]
-        tl.atomic_add(dkv_ptrs_lora, dKV_lora, mask=valid[None, :], sem="relaxed")
-
-        dkv_ptrs_rope = dKV_ptr + safe_pos[None, :] * stride_dkv_t + (D_V + offs_r[:, None])
-        tl.atomic_add(dkv_ptrs_rope, dKV_rope, mask=valid[None, :], sem="relaxed")
-
-
-# =====================================================================
-# Backward method="split_intermediate" — dQ kernel (stores dS/P)
-# =====================================================================
-@triton.jit
-def _bwd_dq_store_intermediates(
-    Q_ptr, KV_ptr, dO_ptr, TopK_ptr, LSE_ptr, Delta_ptr,
-    dQ_ptr,
-    dS_ptr,         # [T, H, TOPK] bf16 -- output
-    P_ptr,          # [T, H, TOPK] bf16 -- output
-    stride_q_t: tl.int64, stride_q_h: tl.int64,
-    stride_kv_t: tl.int64,
-    stride_do_t: tl.int64, stride_do_h: tl.int64,
-    stride_dq_t: tl.int64, stride_dq_h: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_ds_t: tl.int64, stride_ds_h: tl.int64,
-    scale: tl.float32, num_heads: tl.int32,
-    TOPK: tl.constexpr, BLOCK_H: tl.constexpr, TILE_K: tl.constexpr,
-    D_V: tl.constexpr, D_ROPE: tl.constexpr,
-):
-    """dQ kernel that stores dS and P intermediates for the dKV kernel."""
-    token_idx = tl.program_id(0)
-    hg_idx = tl.program_id(1)
-    offs_h = hg_idx * BLOCK_H + tl.arange(0, BLOCK_H)
-    mask_h = offs_h < num_heads
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    q_base = token_idx * stride_q_t
-    Q_lora = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + offs_v[None, :],
-                     mask=mask_h[:, None], other=0.0)
-    Q_rope = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + (D_V + offs_r[None, :]),
-                     mask=mask_h[:, None], other=0.0)
-    do_base = token_idx * stride_do_t
-    dO_val = tl.load(dO_ptr + do_base + offs_h[:, None] * stride_do_h + offs_v[None, :],
-                     mask=mask_h[:, None], other=0.0)
-    lse = tl.load(LSE_ptr + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-    delta = tl.load(Delta_ptr + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-
-    dQ_lora = tl.zeros([BLOCK_H, D_V], dtype=tl.float32)
-    dQ_rope = tl.zeros([BLOCK_H, D_ROPE], dtype=tl.float32)
-    NUM_TILES: tl.constexpr = (TOPK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t
-    offs_tile = tl.arange(0, TILE_K)
-    topk_pos = tl.load(TopK_ptr + topk_base + offs_tile, mask=offs_tile < TOPK, other=-1)
-    topk_pos_next = topk_pos
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        valid = (tile_start + offs_tile) < TOPK
-        valid = valid & (topk_pos != -1)
-        if t + 1 < NUM_TILES:
-            next_offs = (t + 1) * TILE_K + offs_tile
-            topk_pos_next = tl.load(TopK_ptr + topk_base + next_offs,
-                                    mask=next_offs < TOPK, other=-1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        K_lora_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + offs_v[:, None],
-                          mask=valid[None, :], other=0.0)
-        K_rope_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + (D_V + offs_r[:, None]),
-                          mask=valid[None, :], other=0.0)
-
-        S = tl.dot(Q_lora, K_lora_T)
-        S += tl.dot(Q_rope, K_rope_T)
-        S *= scale
-        S = tl.where(valid[None, :] & mask_h[:, None], S, float("-inf"))
-        P = tl.exp(S - lse[:, None])
-        P = tl.where(valid[None, :] & mask_h[:, None], P, 0.0)
-
-        dP = tl.dot(dO_val, K_lora_T)
-        dS = P * (dP - delta[:, None]) * scale
-        dS = tl.where(valid[None, :] & mask_h[:, None], dS, 0.0)
-
-        V_lora = tl.trans(K_lora_T)
-        dQ_lora += tl.dot(dS.to(V_lora.dtype), V_lora).to(tl.float32)
-        K_rope = tl.trans(K_rope_T)
-        dQ_rope += tl.dot(dS.to(K_rope.dtype), K_rope).to(tl.float32)
-
-        # Store dS and P intermediates
-        ds_base = token_idx * stride_ds_t
-        tile_offs = tile_start + offs_tile
-        tl.store(dS_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                 dS.to(tl.bfloat16),
-                 mask=mask_h[:, None] & (tile_offs[None, :] < TOPK))
-        tl.store(P_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                 P.to(tl.bfloat16),
-                 mask=mask_h[:, None] & (tile_offs[None, :] < TOPK))
-
-        if t + 1 < NUM_TILES:
-            topk_pos = topk_pos_next
-
-    dq_base = token_idx * stride_dq_t
-    tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + offs_v[None, :],
-             dQ_lora.to(Q_lora.dtype), mask=mask_h[:, None])
-    tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + (D_V + offs_r[None, :]),
-             dQ_rope.to(Q_rope.dtype), mask=mask_h[:, None])
-
-
-# =====================================================================
-# Backward method="split_intermediate" — dKV kernel (reads dS/P)
-# =====================================================================
-@triton.jit
-def _bwd_dkv_hg_fused(
-    Q_T_ptr,        # [T, D_QK, H] bf16
-    dO_T_ptr,       # [T, D_V, H] bf16
-    dS_ptr,         # [T, H, TOPK] bf16
-    P_ptr,          # [T, H, TOPK] bf16
-    TopK_ptr,       # [T, TOPK] int32
-    dKV_ptr,        # [T, D_QK] fp32
-    stride_qt_t: tl.int64,
-    stride_dot_t: tl.int64,
-    stride_ds_t: tl.int64,
-    stride_ds_h: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_dkv_t: tl.int64,
-    num_heads: tl.int32,
-    TOPK: tl.constexpr,
-    TILE_K: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-    NUM_HG: tl.constexpr,
-    D_V: tl.constexpr,
-    D_ROPE: tl.constexpr,
-):
-    """
-    dKV scatter kernel with head-group fusion.
-
-    Grid: (total_tokens,) -- ONE program per token.
-    Accumulates dKV across all head groups before scattering -- 2x fewer atomics.
-    """
-    token_idx = tl.program_id(0)
-
-    NUM_TILES: tl.constexpr = (TOPK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t
-    offs_tile = tl.arange(0, TILE_K)
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        tile_offs = tile_start + offs_tile
-        topk_pos = tl.load(TopK_ptr + topk_base + tile_offs,
-                           mask=tile_offs < TOPK, other=-1)
-        valid = (tile_offs < TOPK) & (topk_pos != -1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        dKV_lora = tl.zeros([D_V, TILE_K], dtype=tl.float32)
-        dKV_rope = tl.zeros([D_ROPE, TILE_K], dtype=tl.float32)
-
-        for hg in range(NUM_HG):
-            offs_h = hg * BLOCK_H + tl.arange(0, BLOCK_H)
-            mask_h = offs_h < num_heads
-
-            qt_base = token_idx * stride_qt_t
-            Q_lora_T = tl.load(
-                Q_T_ptr + qt_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            Q_rope_T = tl.load(
-                Q_T_ptr + qt_base + (D_V + offs_r[:, None]) * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-
-            dot_base = token_idx * stride_dot_t
-            dO_T = tl.load(
-                dO_T_ptr + dot_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-
-            ds_base = token_idx * stride_ds_t
-            dS_val = tl.load(
-                dS_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & (tile_offs[None, :] < TOPK),
-                other=0.0,
-            )
-            P_val = tl.load(
-                P_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & (tile_offs[None, :] < TOPK),
-                other=0.0,
-            )
-
-            dKV_lora += tl.dot(Q_lora_T, dS_val.to(Q_lora_T.dtype)).to(tl.float32)
-            dKV_lora += tl.dot(dO_T, P_val.to(dO_T.dtype)).to(tl.float32)
-            dKV_rope += tl.dot(Q_rope_T, dS_val.to(Q_rope_T.dtype)).to(tl.float32)
-
-        dkv_ptrs_lora = dKV_ptr + safe_pos[None, :] * stride_dkv_t + offs_v[:, None]
-        tl.atomic_add(dkv_ptrs_lora, dKV_lora, mask=valid[None, :], sem="relaxed")
-
-        dkv_ptrs_rope = dKV_ptr + safe_pos[None, :] * stride_dkv_t + (D_V + offs_r[:, None])
-        tl.atomic_add(dkv_ptrs_rope, dKV_rope, mask=valid[None, :], sem="relaxed")
-
-
-# =====================================================================
-# Backward method="privatized" — dKV kernel with copy-based privatization
-# =====================================================================
-@triton.jit
-def _bwd_dkv_privatized(
-    Q_T_ptr,            # [T, D_QK, H] bf16
-    dO_T_ptr,           # [T, D_V,  H] bf16
-    dS_ptr,             # [T, H, TOPK] bf16
-    P_ptr,              # [T, H, TOPK] bf16
-    TopK_ptr,           # [T, TOPK] int32
-    dKV_copies_ptr,     # [NUM_COPIES, T, D_QK] fp32
-    stride_qt_t: tl.int64,
-    stride_dot_t: tl.int64,
-    stride_ds_t: tl.int64,
-    stride_ds_h: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_copies: tl.int64,   # T * D_QK — gap between consecutive copies
-    stride_dkv_t: tl.int64,   # D_QK
-    num_heads: tl.int32,
-    TOPK: tl.constexpr,
-    TILE_K: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-    NUM_HG: tl.constexpr,
-    NUM_COPIES: tl.constexpr,
-    D_V: tl.constexpr,
-    D_ROPE: tl.constexpr,
-):
-    """
-    Privatized dKV scatter kernel.
-
-    Grid: (total_tokens,) -- one program per query token.
-    Routes atomic_add to copy (token_idx % NUM_COPIES), so each copy
-    receives only TOPK/NUM_COPIES writers per address instead of TOPK.
-    A separate reduction kernel sums the copies afterward.
-    """
-    token_idx = tl.program_id(0)
-    copy_idx = token_idx % NUM_COPIES
-
-    NUM_TILES: tl.constexpr = (TOPK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t
-    offs_tile = tl.arange(0, TILE_K)
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    copy_base = copy_idx * stride_copies
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        tile_offs = tile_start + offs_tile
-        topk_pos = tl.load(TopK_ptr + topk_base + tile_offs,
-                           mask=tile_offs < TOPK, other=-1)
-        valid = (tile_offs < TOPK) & (topk_pos != -1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        dKV_lora = tl.zeros([D_V,    TILE_K], dtype=tl.float32)
-        dKV_rope = tl.zeros([D_ROPE, TILE_K], dtype=tl.float32)
-
-        for hg in range(NUM_HG):
-            offs_h = hg * BLOCK_H + tl.arange(0, BLOCK_H)
-            mask_h = offs_h < num_heads
-
-            qt_base = token_idx * stride_qt_t
-            Q_lora_T = tl.load(
-                Q_T_ptr + qt_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            Q_rope_T = tl.load(
-                Q_T_ptr + qt_base + (D_V + offs_r[:, None]) * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            dot_base = token_idx * stride_dot_t
-            dO_T = tl.load(
-                dO_T_ptr + dot_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            ds_base = token_idx * stride_ds_t
-            dS_val = tl.load(
-                dS_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & (tile_offs[None, :] < TOPK), other=0.0,
-            )
-            P_val = tl.load(
-                P_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & (tile_offs[None, :] < TOPK), other=0.0,
-            )
-
-            dKV_lora += tl.dot(Q_lora_T, dS_val.to(Q_lora_T.dtype)).to(tl.float32)
-            dKV_lora += tl.dot(dO_T,     P_val.to(dO_T.dtype)).to(tl.float32)
-            dKV_rope += tl.dot(Q_rope_T, dS_val.to(Q_rope_T.dtype)).to(tl.float32)
-
-        # Scatter into this token's private copy — NUM_COPIES times fewer writers
-        dkv_lora_ptrs = (dKV_copies_ptr + copy_base
-                         + safe_pos[None, :] * stride_dkv_t + offs_v[:, None])
-        tl.atomic_add(dkv_lora_ptrs, dKV_lora, mask=valid[None, :], sem="relaxed")
-
-        dkv_rope_ptrs = (dKV_copies_ptr + copy_base
-                         + safe_pos[None, :] * stride_dkv_t + (D_V + offs_r[:, None]))
-        tl.atomic_add(dkv_rope_ptrs, dKV_rope, mask=valid[None, :], sem="relaxed")
-
-
-# =====================================================================
-# Backward method="xcd_privatized" — dKV kernel with true XCD-local routing
-# =====================================================================
-@triton.jit
-def _bwd_dkv_xcd_local(
-    Q_T_ptr,            # [T, D_QK, H] bf16
-    dO_T_ptr,           # [T, D_V,  H] bf16
-    dS_ptr,             # [T, H, TOPK] bf16
-    P_ptr,              # [T, H, TOPK] bf16
-    TopK_ptr,           # [T, TOPK] int32
-    dKV_copies_ptr,     # [NUM_XCD, T, D_QK] fp32
-    stride_qt_t: tl.int64,
-    stride_dot_t: tl.int64,
-    stride_ds_t: tl.int64,
-    stride_ds_h: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_copies: tl.int64,   # T * D_QK
-    stride_dkv_t: tl.int64,   # D_QK
-    num_heads: tl.int32,
-    TOPK: tl.constexpr,
-    TILE_K: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-    NUM_HG: tl.constexpr,
-    NUM_XCD: tl.constexpr,
-    D_V: tl.constexpr,
-    D_ROPE: tl.constexpr,
-):
-    """
-    XCD-local privatized dKV scatter kernel.
-
-    Reads the AMDGCN HW_ID hardware register at runtime to obtain the true
-    XCD (shader engine) index for each CTA.  gfx942 bit layout:
-      bits [17:14] = SE_ID = XCD index (0-7 on MI300X).
-    All writers to copy k originate from XCD k — atomic adds stay L2-local,
-    eliminating cross-XCD coherence write-backs.
-    """
-    token_idx = tl.program_id(0)
-    hw_id = tl.inline_asm_elementwise(
-        "s_getreg_b32 s4, hwreg(4, 0, 32)\n"
-        "v_mov_b32 $0, s4",
-        "=v,~{s4}",
-        [],
-        dtype=tl.int32,
-        is_pure=False,
-        pack=1,
-    )
-    copy_idx = (hw_id >> 14) & 0x7  # SE_ID = XCD index, range 0..NUM_XCD-1
-
-    NUM_TILES: tl.constexpr = (TOPK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t
-    offs_tile = tl.arange(0, TILE_K)
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    copy_base = copy_idx * stride_copies
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        tile_offs = tile_start + offs_tile
-        topk_pos = tl.load(TopK_ptr + topk_base + tile_offs,
-                           mask=tile_offs < TOPK, other=-1)
-        valid = (tile_offs < TOPK) & (topk_pos != -1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        dKV_lora = tl.zeros([D_V,    TILE_K], dtype=tl.float32)
-        dKV_rope = tl.zeros([D_ROPE, TILE_K], dtype=tl.float32)
-
-        for hg in range(NUM_HG):
-            offs_h = hg * BLOCK_H + tl.arange(0, BLOCK_H)
-            mask_h = offs_h < num_heads
-
-            qt_base = token_idx * stride_qt_t
-            Q_lora_T = tl.load(
-                Q_T_ptr + qt_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            Q_rope_T = tl.load(
-                Q_T_ptr + qt_base + (D_V + offs_r[:, None]) * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            dot_base = token_idx * stride_dot_t
-            dO_T = tl.load(
-                dO_T_ptr + dot_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            ds_base = token_idx * stride_ds_t
-            dS_val = tl.load(
-                dS_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & (tile_offs[None, :] < TOPK), other=0.0,
-            )
-            P_val = tl.load(
-                P_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & (tile_offs[None, :] < TOPK), other=0.0,
-            )
-
-            dKV_lora += tl.dot(Q_lora_T, dS_val.to(Q_lora_T.dtype)).to(tl.float32)
-            dKV_lora += tl.dot(dO_T,     P_val.to(dO_T.dtype)).to(tl.float32)
-            dKV_rope += tl.dot(Q_rope_T, dS_val.to(Q_rope_T.dtype)).to(tl.float32)
-
-        # Scatter into XCD-local copy — all writers to copy k are on XCD k
-        dkv_lora_ptrs = (dKV_copies_ptr + copy_base
-                         + safe_pos[None, :] * stride_dkv_t + offs_v[:, None])
-        tl.atomic_add(dkv_lora_ptrs, dKV_lora, mask=valid[None, :], sem="relaxed")
-
-        dkv_rope_ptrs = (dKV_copies_ptr + copy_base
-                         + safe_pos[None, :] * stride_dkv_t + (D_V + offs_r[:, None]))
-        tl.atomic_add(dkv_rope_ptrs, dKV_rope, mask=valid[None, :], sem="relaxed")
-
-
-# Experiment: non-atomic scatter — same traffic pattern as _bwd_dkv_hg_fused
-# but uses tl.store instead of tl.atomic_add.  Results are INCORRECT (races),
-# but WRITE_SIZE from rocprof tells us how much of the atomic write traffic is
-# due to the atomic mechanism itself vs. the data volume.
-@triton.jit
-def _bwd_dkv_nonatomic_scatter(
-    Q_T_ptr, dO_T_ptr, dS_ptr, P_ptr, TopK_ptr, dKV_ptr,
-    stride_qt_t: tl.int64, stride_dot_t: tl.int64,
-    stride_ds_t: tl.int64, stride_ds_h: tl.int64,
-    stride_topk_t: tl.int64, stride_dkv_t: tl.int64,
-    num_heads: tl.int32,
-    TOPK: tl.constexpr, TILE_K: tl.constexpr, BLOCK_H: tl.constexpr,
-    NUM_HG: tl.constexpr, D_V: tl.constexpr, D_ROPE: tl.constexpr,
-):
-    """
-    Diagnostic-only: identical compute to _bwd_dkv_hg_fused but writes with
-    tl.store instead of tl.atomic_add.  Results are wrong due to races.
-    Use only for rocprof WRITE_SIZE measurement to isolate atomic overhead.
-    """
-    token_idx = tl.program_id(0)
-    NUM_TILES: tl.constexpr = (TOPK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t
-    offs_tile = tl.arange(0, TILE_K)
-    offs_v    = tl.arange(0, D_V)
-    offs_r    = tl.arange(0, D_ROPE)
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        tile_offs  = tile_start + offs_tile
-        topk_pos   = tl.load(TopK_ptr + topk_base + tile_offs,
-                             mask=tile_offs < TOPK, other=-1)
-        valid   = (tile_offs < TOPK) & (topk_pos != -1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        dKV_lora = tl.zeros([D_V,    TILE_K], dtype=tl.float32)
-        dKV_rope = tl.zeros([D_ROPE, TILE_K], dtype=tl.float32)
-
-        for hg in range(NUM_HG):
-            offs_h = hg * BLOCK_H + tl.arange(0, BLOCK_H)
-            mask_h = offs_h < num_heads
-            qt_base  = token_idx * stride_qt_t
-            dot_base = token_idx * stride_dot_t
-            ds_base  = token_idx * stride_ds_t
-
-            Q_lora_T = tl.load(
-                Q_T_ptr + qt_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0)
-            Q_rope_T = tl.load(
-                Q_T_ptr + qt_base + (D_V + offs_r[:, None]) * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0)
-            dO_T = tl.load(
-                dO_T_ptr + dot_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0)
-            dS_val = tl.load(
-                dS_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & (tile_offs[None, :] < TOPK), other=0.0)
-            P_val = tl.load(
-                P_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & (tile_offs[None, :] < TOPK), other=0.0)
-
-            dKV_lora += tl.dot(Q_lora_T, dS_val.to(Q_lora_T.dtype)).to(tl.float32)
-            dKV_lora += tl.dot(dO_T,     P_val.to(dO_T.dtype)).to(tl.float32)
-            dKV_rope += tl.dot(Q_rope_T, dS_val.to(Q_rope_T.dtype)).to(tl.float32)
-
-        # Non-atomic store — WRONG results, measures write traffic only
-        tl.store(dKV_ptr + safe_pos[None, :] * stride_dkv_t + offs_v[:, None],
-                 dKV_lora, mask=valid[None, :])
-        tl.store(dKV_ptr + safe_pos[None, :] * stride_dkv_t + (D_V + offs_r[:, None]),
-                 dKV_rope, mask=valid[None, :])
-
-
-@triton.jit
-def _bwd_dkv_reduce_copies(
-    dKV_copies_ptr,     # [NUM_COPIES * T * D_QK] fp32, flattened
-    dKV_ptr,            # [T * D_QK] fp32, flattened
-    stride_copies: tl.int64,   # T * D_QK
-    total_elems: tl.int32,
-    NUM_COPIES: tl.constexpr,
-    BLOCK: tl.constexpr,
-):
-    """Sum NUM_COPIES privatized dKV buffers element-wise into dKV."""
-    pid = tl.program_id(0)
-    offs = pid * BLOCK + tl.arange(0, BLOCK)
-    mask = offs < total_elems
-
-    acc = tl.zeros([BLOCK], dtype=tl.float32)
-    for k in tl.static_range(NUM_COPIES):
-        acc += tl.load(dKV_copies_ptr + k * stride_copies + offs,
-                       mask=mask, other=0.0)
-    tl.store(dKV_ptr + offs, acc, mask=mask)
-
-
-# =====================================================================
-# Backward method="gather" — intermediate storage + inverted topk gather
-# =====================================================================
-@triton.jit
-def _bwd_compute_dkv_intermediate(
-    Q_T_ptr,        # [T, D_QK, H] bf16  (q transposed: stride_qt_t = D_QK * H)
-    dO_T_ptr,       # [T, D_V,  H] bf16
-    dS_ptr,         # [T, H, TOPK] bf16
-    P_ptr,          # [T, H, TOPK] bf16
-    TopK_ptr,       # [T, TOPK] int32
-    Interm_ptr,     # [T, TOPK, D_QK] bf16 — output, one writer per (q, topk_rank)
-    stride_qt_t: tl.int64,
-    stride_dot_t: tl.int64,
-    stride_ds_t: tl.int64,
-    stride_ds_h: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_interm_t: tl.int64,   # TOPK * D_QK
-    stride_interm_k: tl.int64,   # D_QK
-    num_heads: tl.int32,
-    TOPK: tl.constexpr,
-    TILE_K: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-    NUM_HG: tl.constexpr,
-    D_V: tl.constexpr,
-    D_ROPE: tl.constexpr,
-):
-    """
-    Same compute as _bwd_dkv_hg_fused but writes to a private intermediate
-    [T, TOPK, D] bf16 instead of atomic_add to shared dKV — no atomics needed.
-
-    Grid: (total_tokens,) -- one program per query token q.
-    For each tile of TOPK, stores dKV_lora/rope for that (q, tile) block.
-    """
-    token_idx = tl.program_id(0)
-
-    NUM_TILES: tl.constexpr = (TOPK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t
-    offs_tile = tl.arange(0, TILE_K)
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    interm_base_t = token_idx * stride_interm_t  # base for this query token
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        tile_offs = tile_start + offs_tile
-        valid = tile_offs < TOPK
-
-        dKV_lora = tl.zeros([D_V, TILE_K], dtype=tl.float32)
-        dKV_rope = tl.zeros([D_ROPE, TILE_K], dtype=tl.float32)
-
-        for hg in range(NUM_HG):
-            offs_h = hg * BLOCK_H + tl.arange(0, BLOCK_H)
-            mask_h = offs_h < num_heads
-
-            qt_base = token_idx * stride_qt_t
-            Q_lora_T = tl.load(
-                Q_T_ptr + qt_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-            Q_rope_T = tl.load(
-                Q_T_ptr + qt_base + (D_V + offs_r[:, None]) * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-
-            dot_base = token_idx * stride_dot_t
-            dO_T = tl.load(
-                dO_T_ptr + dot_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                mask=mask_h[None, :], other=0.0,
-            )
-
-            ds_base = token_idx * stride_ds_t
-            dS_val = tl.load(
-                dS_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & valid[None, :], other=0.0,
-            )
-            P_val = tl.load(
-                P_ptr + ds_base + offs_h[:, None] * stride_ds_h + tile_offs[None, :],
-                mask=mask_h[:, None] & valid[None, :], other=0.0,
-            )
-
-            dKV_lora += tl.dot(Q_lora_T, dS_val.to(Q_lora_T.dtype)).to(tl.float32)
-            dKV_lora += tl.dot(dO_T, P_val.to(dO_T.dtype)).to(tl.float32)
-            dKV_rope += tl.dot(Q_rope_T, dS_val.to(Q_rope_T.dtype)).to(tl.float32)
-
-        # Store to intermediate: Interm[token_idx, tile_start:tile_start+TILE_K, 0:D_V]
-        # Layout: [T, TOPK, D] so pointer = interm_base_t + tile_offs[None,:]*D + offs_v[:,None]
-        interm_lora_ptrs = (Interm_ptr + interm_base_t
-                            + tile_offs[None, :] * stride_interm_k
-                            + offs_v[:, None])
-        tl.store(interm_lora_ptrs, dKV_lora.to(tl.bfloat16), mask=valid[None, :])
-
-        interm_rope_ptrs = (Interm_ptr + interm_base_t
-                            + tile_offs[None, :] * stride_interm_k
-                            + D_V + offs_r[:, None])
-        tl.store(interm_rope_ptrs, dKV_rope.to(tl.bfloat16), mask=valid[None, :])
-
-
-@triton.jit
-def _bwd_dkv_gather(
-    Interm_ptr,     # [T, TOPK, D] bf16, flattened as [T*TOPK, D]
-    InvPtr_ptr,     # [T+1] int32 — CSR row pointers (kv_token -> range in inv_data)
-    InvData_ptr,    # [T*TOPK] int32 — encoded as q*TOPK+r, sorted by KV token
-    dKV_ptr,        # [T, D] bf16 — output
-    stride_interm_k: tl.int64,   # D_V + D_ROPE
-    stride_dkv_t: tl.int64,
-    TOPK: tl.constexpr,
-    D_V: tl.constexpr,
-    D_ROPE: tl.constexpr,
-):
-    """
-    Gather dKV from intermediate buffer using CSR-style inverted topk index.
-
-    Grid: (total_tokens,) -- one CTA per KV token k.
-    Accumulates in fp32, stores bf16. No atomics.
-    """
-    k = tl.program_id(0)
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    start = tl.load(InvPtr_ptr + k)
-    end = tl.load(InvPtr_ptr + k + 1)
-
-    dkv_acc_lora = tl.zeros([D_V], dtype=tl.float32)
-    dkv_acc_rope = tl.zeros([D_ROPE], dtype=tl.float32)
-
-    n_entries = end - start
-    for i in range(n_entries):
-        # entry = q*TOPK + r, used directly as flat index into [T*TOPK, D] intermediate
-        entry = tl.load(InvData_ptr + start + i).to(tl.int64)
-        base = entry * stride_interm_k
-        lora_val = tl.load(Interm_ptr + base + offs_v)
-        rope_val = tl.load(Interm_ptr + base + D_V + offs_r)
-        dkv_acc_lora += lora_val.to(tl.float32)
-        dkv_acc_rope += rope_val.to(tl.float32)
-
-    dkv_base = k.to(tl.int64) * stride_dkv_t
-    tl.store(dKV_ptr + dkv_base + offs_v, dkv_acc_lora.to(tl.bfloat16))
-    tl.store(dKV_ptr + dkv_base + D_V + offs_r, dkv_acc_rope.to(tl.bfloat16))
-
-
-def _build_inverted_topk(topk_indices):
-    """
-    Build CSR-style inverted index from topk_indices [T, TOPK] int32.
-
-    Returns:
-        inv_ptr:  [T+1] int32 — row pointers (kv_token -> range in inv_data)
-        inv_data: [T*TOPK] int32 — encoded (q*TOPK+r) values, sorted by KV token
-    """
-    T, TOPK = topk_indices.shape
-    device = topk_indices.device
-
-    flat_kv = topk_indices.reshape(-1).long()   # [T*TOPK] KV token indices
-    # argsort by KV token to get the flat indices (q*TOPK+r) in sorted order
-    order = torch.argsort(flat_kv, stable=True)
-    inv_data = order.to(torch.int32)  # [T*TOPK]
-
-    counts = torch.zeros(T, dtype=torch.int32, device=device)
-    counts.scatter_add_(0, flat_kv, torch.ones(T * TOPK, dtype=torch.int32, device=device))
-
-    inv_ptr = torch.zeros(T + 1, dtype=torch.int32, device=device)
-    torch.cumsum(counts, dim=0, out=inv_ptr[1:])
-
-    return inv_ptr, inv_data
-
-
-def _build_inverted_topk_slice(topk_indices_slice, r_start, R_CHUNK):
-    """
-    Build CSR-style inverted index for a topk slice, excluding invalid (-1) entries.
-
-    Args:
-        topk_indices_slice: [T, R_CHUNK] int32 — topk_indices[:, r_start:r_start+R_CHUNK]
-          May contain -1 for padding (when actual chunk < R_CHUNK at the last chunk).
-        r_start:  int — first rank index in this slice (unused, for documentation)
-        R_CHUNK:  int — number of ranks in this slice (constexpr width)
-
-    Returns:
-        inv_ptr:  [T+1] int32 — row pointers (kv_token -> range in inv_data)
-        inv_data: [valid_entries] int32 — flat indices q*R_CHUNK+local_r, sorted by KV token
-    """
-    T, RC = topk_indices_slice.shape
-    device = topk_indices_slice.device
-
-    flat_idx = torch.arange(T * RC, dtype=torch.int32, device=device)  # q*RC + local_r
-    flat_kv = topk_indices_slice.reshape(-1).long()   # [T*R_CHUNK] KV token indices
-
-    # Exclude invalid entries (kv_token == -1)
-    valid_mask = flat_kv >= 0
-    valid_flat_kv = flat_kv[valid_mask]
-    valid_flat_idx = flat_idx[valid_mask]
-
-    # Sort valid entries by KV token
-    order = torch.argsort(valid_flat_kv, stable=True)
-    inv_data = valid_flat_idx[order].to(torch.int32)
-
-    counts = torch.zeros(T, dtype=torch.int32, device=device)
-    counts.scatter_add_(0, valid_flat_kv,
-                        torch.ones(valid_flat_kv.shape[0], dtype=torch.int32, device=device))
-
-    inv_ptr = torch.zeros(T + 1, dtype=torch.int32, device=device)
-    torch.cumsum(counts, dim=0, out=inv_ptr[1:])
-
-    return inv_ptr, inv_data
-
-
-# =====================================================================
-# Backward method="persistent" — Approach A: persistent 304-CTA kernel
-# =====================================================================
-@triton.jit
-def _bwd_persistent_chunk(
-    Q_ptr,          # [T, H, D] bf16
-    KV_ptr,         # [T, 1, D] bf16
-    dO_ptr,         # [T, H, D_V] bf16
-    TopK_ptr,       # [T, TOPK] int32
-    LSE_ptr,        # [T, H] fp32
-    Delta_ptr,      # [T, H] fp32
-    dQ_ptr,         # [T, H, D] bf16 — output
-    dKV_chunk_ptr,  # [NUM_XCD, K_CHUNK, D] fp32 — XCD-local accumulator
-    stride_q_t: tl.int64, stride_q_h: tl.int64,
-    stride_kv_t: tl.int64,
-    stride_do_t: tl.int64, stride_do_h: tl.int64,
-    stride_dq_t: tl.int64, stride_dq_h: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_chunk_xcd: tl.int64,   # K_CHUNK * D
-    stride_chunk_k: tl.int64,     # D
-    scale: tl.float32,
-    total_tokens: tl.int32,
-    num_heads: tl.int32,
-    k_start: tl.int32,            # first KV token index in this chunk
-    k_end: tl.int32,              # one past last KV token index in this chunk
-    TOPK: tl.int32,               # NOT constexpr — avoids unrolling 1024 iterations
-    TOKENS_PER_CU: tl.int32,     # NOT constexpr — avoids unrolling 14 token iterations
-    BLOCK_H: tl.constexpr,
-    NUM_HG: tl.constexpr,
-    D_V: tl.constexpr,
-    D_ROPE: tl.constexpr,
-):
-    """
-    Persistent backward kernel for Approach A.
-
-    grid=(NUM_CUS,) — one CTA per physical CU on MI300X.
-    Each CTA:
-      1. Reads hw_id to obtain its XCD index (0-7) for L2-local atomics.
-      2. Owns tokens [pid*(T//NUM_CUS), (pid+1)*(T//NUM_CUS)).
-      3. For each owned token: loops all TOPK ranks, accumulates dQ in fp32
-         registers, atomic_adds dKV contributions into dKV_chunk[xcd, k-k_start, :]
-         (only for KV tokens in [k_start, k_end)).
-      4. Writes dQ[q] once per token.
-
-    dKV_chunk[xcd] is K_CHUNK*D fp32 = 3.14 MB for K_CHUNK=1365, D=576.
-    This fits in the 4 MB L2 per XCD so atomics stay L2-local.
-    """
-    pid = tl.program_id(0)
-
-    # XCD index: use pid % 8 as a uniform proxy for XCD assignment.
-    # (Reading hw_id via inline_asm_elementwise returns a vgpr/per-lane tensor
-    #  which causes Triton to generate per-lane scatter addressing, blowing up
-    #  compilation. The static formula is sufficient since the kernel's benefit
-    #  comes from K_CHUNK fitting in L2, not from precise CTA→XCD routing.)
-    xcd = pid % 8
-
-    # Token range owned by this CTA — last CTA may own fewer tokens (guarded by valid_q)
-    q_start = pid * TOKENS_PER_CU
-
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-    chunk_base = xcd.to(tl.int64) * stride_chunk_xcd
-
-    for q_off in range(TOKENS_PER_CU):
-        q = q_start + q_off
-        valid_q = q < total_tokens  # guard for last CTA which may own fewer tokens
-
-        q_base  = q * stride_q_t
-        do_base = q * stride_do_t
-        topk_base = q * stride_topk_t
-
-        for hg in range(NUM_HG):
-            offs_h = hg * BLOCK_H + tl.arange(0, BLOCK_H)
-            mask_h = (offs_h < num_heads) & valid_q
-
-            Q_lora = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + offs_v[None, :],
-                             mask=mask_h[:, None], other=0.0)   # [BLOCK_H, D_V]
-            Q_rope = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + (D_V + offs_r[None, :]),
-                             mask=mask_h[:, None], other=0.0)   # [BLOCK_H, D_ROPE]
-            dO_val = tl.load(dO_ptr + do_base + offs_h[:, None] * stride_do_h + offs_v[None, :],
-                             mask=mask_h[:, None], other=0.0)   # [BLOCK_H, D_V]
-            lse   = tl.load(LSE_ptr   + q * num_heads + offs_h, mask=mask_h, other=0.0)
-            delta = tl.load(Delta_ptr + q * num_heads + offs_h, mask=mask_h, other=0.0)
-
-            dQ_lora = tl.zeros([BLOCK_H, D_V],   dtype=tl.float32)
-            dQ_rope = tl.zeros([BLOCK_H, D_ROPE], dtype=tl.float32)
-
-            for r in range(TOPK):
-                kv_pos = tl.load(TopK_ptr + topk_base + r, mask=valid_q, other=0).to(tl.int32)
-
-                # Load KV token and promote to fp32 for accurate dot products
-                K_lora_T = tl.load(KV_ptr + kv_pos * stride_kv_t + offs_v).to(tl.float32)    # [D_V]
-                K_rope_T = tl.load(KV_ptr + kv_pos * stride_kv_t + D_V + offs_r).to(tl.float32)  # [D_ROPE]
-
-                # Attention scores S = Q @ K^T → [BLOCK_H] (fp32 accumulation)
-                S = tl.sum(Q_lora.to(tl.float32) * K_lora_T[None, :], axis=1) \
-                  + tl.sum(Q_rope.to(tl.float32) * K_rope_T[None, :], axis=1)
-                S = tl.where(mask_h, S * scale, float("-inf"))
-                P = tl.exp(S - lse)
-                P = tl.where(mask_h, P, 0.0)
-
-                # dS = P * (dO @ K^T - delta) * scale → [BLOCK_H]
-                dP = tl.sum(dO_val.to(tl.float32) * K_lora_T[None, :], axis=1)
-                dS = P * (dP - delta) * scale
-                dS = tl.where(mask_h, dS, 0.0)
-
-                # Accumulate dQ (fp32)
-                dQ_lora += dS[:, None] * K_lora_T[None, :]
-                dQ_rope += dS[:, None] * K_rope_T[None, :]
-
-                # Accumulate dKV for this chunk — only for KV tokens in [k_start, k_end)
-                in_chunk = (kv_pos >= k_start) & (kv_pos < k_end)
-                local_k = tl.where(in_chunk, kv_pos - k_start, 0)
-                dkv_ptr_lora = dKV_chunk_ptr + chunk_base + local_k * stride_chunk_k + offs_v
-                dkv_ptr_rope = dKV_chunk_ptr + chunk_base + local_k * stride_chunk_k + D_V + offs_r
-                # dKV_lora += dS * Q_lora + P * dO  (summed over BLOCK_H heads, fp32)
-                dkv_contrib_lora = tl.sum(dS[:, None] * Q_lora.to(tl.float32), axis=0) \
-                                 + tl.sum(P[:, None]  * dO_val.to(tl.float32), axis=0)
-                dkv_contrib_rope = tl.sum(dS[:, None] * Q_rope.to(tl.float32), axis=0)
-                # Broadcast in_chunk scalar to vector mask — suppresses add for out-of-chunk tokens
-                mask_v = in_chunk & (offs_v < D_V)
-                mask_r = in_chunk & (offs_r < D_ROPE)
-                tl.atomic_add(dkv_ptr_lora, dkv_contrib_lora, mask=mask_v)
-                tl.atomic_add(dkv_ptr_rope, dkv_contrib_rope, mask=mask_r)
-
-            # Write dQ for this head group
-            dq_base = q * stride_dq_t
-            tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + offs_v[None, :],
-                     dQ_lora.to(Q_lora.dtype), mask=mask_h[:, None])
-            tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + (D_V + offs_r[None, :]),
-                     dQ_rope.to(Q_rope.dtype), mask=mask_h[:, None])
-
-
-@triton.jit
-def _bwd_chunk_reduce(
-    dKV_chunk_ptr,  # [NUM_XCD, K_CHUNK, D] fp32
-    dKV_ptr,        # [T, D] fp32 — output
-    stride_chunk_xcd: tl.int64,
-    stride_dkv_t: tl.int64,
-    k_start: tl.int32,
-    K_CHUNK: tl.constexpr,
-    NUM_XCD: tl.constexpr,
-    D: tl.constexpr,
-    BLOCK_D: tl.constexpr,
-):
-    """
-    Reduce the NUM_XCD dkv_chunk copies for one K_CHUNK into dKV[k_start:k_end].
-    Grid: (K_CHUNK, D // BLOCK_D)
-    """
-    k_local = tl.program_id(0)
-    d_block = tl.program_id(1)
-    offs_d  = d_block * BLOCK_D + tl.arange(0, BLOCK_D)
-    mask_d  = offs_d < D
-
-    acc = tl.zeros([BLOCK_D], dtype=tl.float32)
-    for xcd in tl.static_range(NUM_XCD):
-        ptr = dKV_chunk_ptr + xcd * stride_chunk_xcd + k_local * D + offs_d
-        acc += tl.load(ptr, mask=mask_d, other=0.0)
-
-    kv_token = k_start + k_local
-    tl.store(dKV_ptr + kv_token * stride_dkv_t + offs_d, acc, mask=mask_d)
-
-
-# =====================================================================
-# Backward method="chunked_gather" — three kernels per chunk (no atomics)
-# =====================================================================
-@triton.jit
-def _bwd_chunk_dq_store_ds(
-    Q_ptr,          # [T, H, D] bf16
-    KV_ptr,         # [T, 1, D] bf16
-    dO_ptr,         # [T, H, D_V] bf16
-    TopK_ptr,       # [T, TOPK] int32
-    LSE_ptr,        # [T, H] fp32
-    Delta_ptr,      # [T, H] fp32
-    dQ_ptr,         # [T, H, D] bf16 — read-modify-write across chunks
-    dS_ptr,         # [T, H, R_CHUNK] bf16 — output chunk dS
-    P_ptr,          # [T, H, R_CHUNK] bf16 — output chunk P
-    stride_q_t: tl.int64, stride_q_h: tl.int64,
-    stride_kv_t: tl.int64,
-    stride_do_t: tl.int64, stride_do_h: tl.int64,
-    stride_dq_t: tl.int64, stride_dq_h: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_ds_t: tl.int64, stride_ds_h: tl.int64,
-    scale: tl.float32, num_heads: tl.int32,
-    R_START: tl.int32,
-    R_CHUNK: tl.constexpr,
-    BLOCK_H: tl.constexpr, TILE_K: tl.constexpr,
-    D_V: tl.constexpr, D_ROPE: tl.constexpr,
-    IS_FIRST_CHUNK: tl.constexpr,
-):
-    """
-    dQ accumulation for rank chunk [R_START, R_START+R_CHUNK), plus stores
-    chunk dS and P to [T, H, R_CHUNK] buffers for use by _bwd_compute_dkv_intermediate.
-    Grid: (total_tokens, num_hg).
-    """
-    token_idx = tl.program_id(0)
-    hg_idx = tl.program_id(1)
-    offs_h = hg_idx * BLOCK_H + tl.arange(0, BLOCK_H)
-    mask_h = offs_h < num_heads
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    q_base = token_idx * stride_q_t
-    Q_lora = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + offs_v[None, :],
-                     mask=mask_h[:, None], other=0.0)
-    Q_rope = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + (D_V + offs_r[None, :]),
-                     mask=mask_h[:, None], other=0.0)
-    do_base = token_idx * stride_do_t
-    dO_val = tl.load(dO_ptr + do_base + offs_h[:, None] * stride_do_h + offs_v[None, :],
-                     mask=mask_h[:, None], other=0.0)
-    lse   = tl.load(LSE_ptr   + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-    delta = tl.load(Delta_ptr + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-
-    dq_base = token_idx * stride_dq_t
-    if IS_FIRST_CHUNK:
-        dQ_lora = tl.zeros([BLOCK_H, D_V],   dtype=tl.float32)
-        dQ_rope = tl.zeros([BLOCK_H, D_ROPE], dtype=tl.float32)
-    else:
-        dQ_lora = tl.load(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + offs_v[None, :],
-                          mask=mask_h[:, None], other=0.0).to(tl.float32)
-        dQ_rope = tl.load(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + (D_V + offs_r[None, :]),
-                          mask=mask_h[:, None], other=0.0).to(tl.float32)
-
-    NUM_TILES: tl.constexpr = (R_CHUNK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t + R_START
-    offs_tile = tl.arange(0, TILE_K)
-    ds_base = token_idx * stride_ds_t + hg_idx * BLOCK_H * stride_ds_h
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        tile_offs  = tile_start + offs_tile
-        valid = tile_offs < R_CHUNK
-        topk_pos = tl.load(TopK_ptr + topk_base + tile_offs, mask=valid, other=-1)
-        valid = valid & (topk_pos != -1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        K_lora_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + offs_v[:, None],
-                           mask=valid[None, :], other=0.0)
-        K_rope_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + (D_V + offs_r[:, None]),
-                           mask=valid[None, :], other=0.0)
-
-        S  = tl.dot(Q_lora, K_lora_T) + tl.dot(Q_rope, K_rope_T)
-        S  = tl.where(valid[None, :] & mask_h[:, None], S * scale, float("-inf"))
-        P  = tl.exp(S - lse[:, None])
-        P  = tl.where(valid[None, :] & mask_h[:, None], P, 0.0)
-        dP = tl.dot(dO_val, K_lora_T)
-        dS = P * (dP - delta[:, None]) * scale
-        dS = tl.where(valid[None, :] & mask_h[:, None], dS, 0.0)
-
-        dQ_lora += tl.dot(dS.to(tl.bfloat16), tl.trans(K_lora_T)).to(tl.float32)
-        dQ_rope += tl.dot(dS.to(tl.bfloat16), tl.trans(K_rope_T)).to(tl.float32)
-
-        # Store chunk dS and P for this tile — use local head offsets (0..BLOCK_H-1)
-        # since ds_base already encodes hg_idx*BLOCK_H*stride_ds_h
-        local_h = tl.arange(0, BLOCK_H)
-        tl.store(dS_ptr + ds_base + local_h[:, None] * stride_ds_h + tile_offs[None, :],
-                 dS.to(tl.bfloat16), mask=mask_h[:, None] & valid[None, :])
-        tl.store(P_ptr  + ds_base + local_h[:, None] * stride_ds_h + tile_offs[None, :],
-                 P.to(tl.bfloat16),  mask=mask_h[:, None] & valid[None, :])
-
-    tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + offs_v[None, :],
-             dQ_lora.to(Q_lora.dtype), mask=mask_h[:, None])
-    tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + (D_V + offs_r[None, :]),
-             dQ_rope.to(Q_rope.dtype), mask=mask_h[:, None])
-
-
-@triton.jit
-def _bwd_chunk_dq(
-    Q_ptr,          # [T, H, D] bf16
-    KV_ptr,         # [T, 1, D] bf16
-    dO_ptr,         # [T, H, D_V] bf16
-    TopK_ptr,       # [T, TOPK] int32
-    LSE_ptr,        # [T, H] fp32
-    Delta_ptr,      # [T, H] fp32
-    dQ_ptr,         # [T, H, D] bf16 — read-modify-write across chunks
-    stride_q_t: tl.int64, stride_q_h: tl.int64,
-    stride_kv_t: tl.int64,
-    stride_do_t: tl.int64, stride_do_h: tl.int64,
-    stride_dq_t: tl.int64, stride_dq_h: tl.int64,
-    stride_topk_t: tl.int64,
-    scale: tl.float32, num_heads: tl.int32,
-    R_START: tl.int32,
-    R_CHUNK: tl.constexpr,
-    BLOCK_H: tl.constexpr, TILE_K: tl.constexpr,
-    D_V: tl.constexpr, D_ROPE: tl.constexpr,
-    IS_FIRST_CHUNK: tl.constexpr,
-):
-    """
-    dQ accumulation for one rank chunk [R_START, R_START+R_CHUNK).
-    Grid: (total_tokens, num_hg). No writes to intermediate buffer.
-    IS_FIRST_CHUNK=True: initialises dQ to zero (avoids a global memory read).
-    """
-    token_idx = tl.program_id(0)
-    hg_idx = tl.program_id(1)
-    offs_h = hg_idx * BLOCK_H + tl.arange(0, BLOCK_H)
-    mask_h = offs_h < num_heads
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    q_base = token_idx * stride_q_t
-    Q_lora = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + offs_v[None, :],
-                     mask=mask_h[:, None], other=0.0)
-    Q_rope = tl.load(Q_ptr + q_base + offs_h[:, None] * stride_q_h + (D_V + offs_r[None, :]),
-                     mask=mask_h[:, None], other=0.0)
-    do_base = token_idx * stride_do_t
-    dO_val = tl.load(dO_ptr + do_base + offs_h[:, None] * stride_do_h + offs_v[None, :],
-                     mask=mask_h[:, None], other=0.0)
-    lse   = tl.load(LSE_ptr   + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-    delta = tl.load(Delta_ptr + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-
-    dq_base = token_idx * stride_dq_t
-    if IS_FIRST_CHUNK:
-        dQ_lora = tl.zeros([BLOCK_H, D_V],   dtype=tl.float32)
-        dQ_rope = tl.zeros([BLOCK_H, D_ROPE], dtype=tl.float32)
-    else:
-        dQ_lora = tl.load(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + offs_v[None, :],
-                          mask=mask_h[:, None], other=0.0).to(tl.float32)
-        dQ_rope = tl.load(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + (D_V + offs_r[None, :]),
-                          mask=mask_h[:, None], other=0.0).to(tl.float32)
-
-    NUM_TILES: tl.constexpr = (R_CHUNK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t + R_START
-    offs_tile = tl.arange(0, TILE_K)
-    topk_pos = tl.load(TopK_ptr + topk_base + offs_tile, mask=offs_tile < R_CHUNK, other=-1)
-    topk_pos_next = topk_pos
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        valid = (tile_start + offs_tile) < R_CHUNK
-        valid = valid & (topk_pos != -1)
-        if t + 1 < NUM_TILES:
-            next_offs = (t + 1) * TILE_K + offs_tile
-            topk_pos_next = tl.load(TopK_ptr + topk_base + next_offs,
-                                    mask=next_offs < R_CHUNK, other=-1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        K_lora_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + offs_v[:, None],
-                           mask=valid[None, :], other=0.0)
-        K_rope_T = tl.load(KV_ptr + safe_pos[None, :] * stride_kv_t + (D_V + offs_r[:, None]),
-                           mask=valid[None, :], other=0.0)
-
-        S  = tl.dot(Q_lora, K_lora_T) + tl.dot(Q_rope, K_rope_T)
-        S  = tl.where(valid[None, :] & mask_h[:, None], S * scale, float("-inf"))
-        P  = tl.exp(S - lse[:, None])
-        P  = tl.where(valid[None, :] & mask_h[:, None], P, 0.0)
-        dP = tl.dot(dO_val, K_lora_T)
-        dS = P * (dP - delta[:, None]) * scale
-        dS = tl.where(valid[None, :] & mask_h[:, None], dS, 0.0)
-
-        dQ_lora += tl.dot(dS.to(tl.bfloat16), tl.trans(K_lora_T)).to(tl.float32)
-        dQ_rope += tl.dot(dS.to(tl.bfloat16), tl.trans(K_rope_T)).to(tl.float32)
-
-        if t + 1 < NUM_TILES:
-            topk_pos = topk_pos_next
-
-    tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + offs_v[None, :],
-             dQ_lora.to(Q_lora.dtype), mask=mask_h[:, None])
-    tl.store(dQ_ptr + dq_base + offs_h[:, None] * stride_dq_h + (D_V + offs_r[None, :]),
-             dQ_rope.to(Q_rope.dtype), mask=mask_h[:, None])
-
-
-@triton.jit
-def _bwd_chunk_dkv_interm(
-    Q_T_ptr,        # [T, D_QK, H] bf16  (transposed: stride = D_QK * H)
-    dO_T_ptr,       # [T, D_V,  H] bf16
-    TopK_ptr,       # [T, TOPK] int32
-    LSE_ptr,        # [T, H] fp32
-    Delta_ptr,      # [T, H] fp32
-    KV_ptr,         # [T, 1, D] bf16
-    Interm_ptr,     # [T, R_CHUNK, D] bf16 — output (plain store, one writer per slot)
-    stride_qt_t: tl.int64,
-    stride_dot_t: tl.int64,
-    stride_topk_t: tl.int64,
-    stride_kv_t: tl.int64,
-    stride_interm_t: tl.int64,   # R_CHUNK * D
-    stride_interm_r: tl.int64,   # D
-    scale: tl.float32, num_heads: tl.int32,
-    R_START: tl.int32,
-    R_CHUNK: tl.constexpr,
-    TILE_K: tl.constexpr,
-    BLOCK_H: tl.constexpr,
-    NUM_HG: tl.constexpr,
-    D_V: tl.constexpr, D_ROPE: tl.constexpr,
-):
-    """
-    dKV intermediate for one rank chunk [R_START, R_START+R_CHUNK).
-    Grid: (total_tokens,) — one CTA per query token, inner loop over head groups.
-    Recomputes S/P/dS on-the-fly. Plain stores to bf16 interm — no atomics.
-    """
-    token_idx = tl.program_id(0)
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-    offs_tile = tl.arange(0, TILE_K)
-
-    NUM_TILES: tl.constexpr = (R_CHUNK + TILE_K - 1) // TILE_K
-    topk_base = token_idx * stride_topk_t + R_START
-    interm_base_t = token_idx * stride_interm_t
-
-    qt_base  = token_idx * stride_qt_t
-    dot_base = token_idx * stride_dot_t
-
-    for t in range(NUM_TILES):
-        tile_start = t * TILE_K
-        tile_offs  = tile_start + offs_tile
-        valid_tile = tile_offs < R_CHUNK
-
-        topk_pos = tl.load(TopK_ptr + topk_base + tile_start + offs_tile,
-                           mask=valid_tile, other=-1)
-        valid    = valid_tile & (topk_pos != -1)
-        safe_pos = tl.where(valid, topk_pos, 0)
-
-        K_lora = tl.load(KV_ptr + safe_pos[:, None] * stride_kv_t + offs_v[None, :],
-                         mask=valid[:, None], other=0.0)   # [TILE_K, D_V]
-        K_rope = tl.load(KV_ptr + safe_pos[:, None] * stride_kv_t + (D_V + offs_r[None, :]),
-                         mask=valid[:, None], other=0.0)   # [TILE_K, D_ROPE]
-
-        dKV_lora = tl.zeros([TILE_K, D_V],   dtype=tl.float32)
-        dKV_rope = tl.zeros([TILE_K, D_ROPE], dtype=tl.float32)
-
-        for hg in range(NUM_HG):
-            offs_h = hg * BLOCK_H + tl.arange(0, BLOCK_H)
-            mask_h = offs_h < num_heads
-
-            Q_lora_T = tl.load(Q_T_ptr + qt_base  + offs_v[:, None] * num_heads + offs_h[None, :],
-                                mask=mask_h[None, :], other=0.0)          # [D_V, BLOCK_H]
-            Q_rope_T = tl.load(Q_T_ptr + qt_base  + (D_V + offs_r[:, None]) * num_heads + offs_h[None, :],
-                                mask=mask_h[None, :], other=0.0)          # [D_ROPE, BLOCK_H]
-            dO_T     = tl.load(dO_T_ptr + dot_base + offs_v[:, None] * num_heads + offs_h[None, :],
-                                mask=mask_h[None, :], other=0.0)          # [D_V, BLOCK_H]
-            lse_h    = tl.load(LSE_ptr   + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-            delta_h  = tl.load(Delta_ptr + token_idx * num_heads + offs_h, mask=mask_h, other=0.0)
-
-            # S = K @ Q^T  [TILE_K, BLOCK_H]
-            S  = tl.dot(K_lora, Q_lora_T) + tl.dot(K_rope, Q_rope_T)
-            S  = tl.where(valid[:, None] & mask_h[None, :], S * scale, float("-inf"))
-            P  = tl.exp(S - lse_h[None, :])
-            P  = tl.where(valid[:, None] & mask_h[None, :], P, 0.0)
-            dP = tl.dot(K_lora, dO_T)   # [TILE_K, BLOCK_H]
-            dS = P * (dP - delta_h[None, :]) * scale
-            dS = tl.where(valid[:, None] & mask_h[None, :], dS, 0.0)
-
-            dKV_lora += tl.dot(dS.to(tl.bfloat16), tl.trans(Q_lora_T)).to(tl.float32)
-            dKV_lora += tl.dot(P.to(tl.bfloat16),  tl.trans(dO_T)).to(tl.float32)
-            dKV_rope += tl.dot(dS.to(tl.bfloat16), tl.trans(Q_rope_T)).to(tl.float32)
-
-        # Plain store to bf16 interm — one writer per (token, local_r) slot
-        interm_lora_ptrs = (Interm_ptr + interm_base_t
-                            + tile_offs[:, None] * stride_interm_r + offs_v[None, :])
-        tl.store(interm_lora_ptrs, dKV_lora.to(tl.bfloat16), mask=valid[:, None])
-
-        interm_rope_ptrs = (Interm_ptr + interm_base_t
-                            + tile_offs[:, None] * stride_interm_r + (D_V + offs_r[None, :]))
-        tl.store(interm_rope_ptrs, dKV_rope.to(tl.bfloat16), mask=valid[:, None])
-
-
-@triton.jit
-def _bwd_dkv_gather_acc(
-    Interm_ptr,     # [T, R_CHUNK, D] bf16 — chunk intermediate
-    InvPtr_ptr,     # [T+1] int32 — CSR row pointers
-    InvData_ptr,    # [T*R_CHUNK] int32 — encoded as q*R_CHUNK + local_r
-    dKV_acc_ptr,    # [T, D] fp32 — accumulator (read-modify-write across chunks)
-    stride_interm_r: tl.int64,   # D
-    stride_acc_t: tl.int64,      # D
-    D_V: tl.constexpr,
-    D_ROPE: tl.constexpr,
-):
-    """
-    Gather one chunk's bf16 intermediate into the fp32 dKV accumulator.
-    Grid: (total_tokens,) — one CTA per KV token k, no atomics.
-    """
-    k = tl.program_id(0)
-    offs_v = tl.arange(0, D_V)
-    offs_r = tl.arange(0, D_ROPE)
-
-    start = tl.load(InvPtr_ptr + k)
-    end   = tl.load(InvPtr_ptr + k + 1)
-
-    acc_base = k.to(tl.int64) * stride_acc_t
-    dkv_acc_lora = tl.load(dKV_acc_ptr + acc_base + offs_v).to(tl.float32)
-    dkv_acc_rope = tl.load(dKV_acc_ptr + acc_base + D_V + offs_r).to(tl.float32)
-
-    n_entries = end - start
-    for i in range(n_entries):
-        entry = tl.load(InvData_ptr + start + i).to(tl.int64)
-        base  = entry * stride_interm_r
-        dkv_acc_lora += tl.load(Interm_ptr + base + offs_v).to(tl.float32)
-        dkv_acc_rope += tl.load(Interm_ptr + base + D_V + offs_r).to(tl.float32)
-
-    tl.store(dKV_acc_ptr + acc_base + offs_v,       dkv_acc_lora)
-    tl.store(dKV_acc_ptr + acc_base + D_V + offs_r, dkv_acc_rope)
 
 
 # =====================================================================
